# src/main/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: Video Game Library API
  version: 1.0.0
  description: API para gestionar una biblioteca de videojuegos (integración con IGDB)
  contact:
    name: Francisco José Soler Conchello
    email: franciscosoler.dev@gmail.com
    url: https://www.linkedin.com/in/franciscojosesoler/
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
tags:
  - name: Games
    description: Operaciones relacionadas con videojuegos
  - name: Platforms
    description: Operaciones relacionadas con plataformas de videojuegos
  - name: Users
    description: Operaciones relacionadas con usuarios
  - name: Library
    description: Operaciones relacionadas con la biblioteca de juegos de un usuario

security: # Seguridad global para todos los endpoints (excepto los que tienen 'permitAll')
  - bearerAuth: []

paths:
  /games/search:
    get:
      tags:
        - Games
      summary: Buscar videojuegos por nombre
      description: >
        Busca videojuegos en la API externa de IGDB por su nombre.
        Devuelve una lista de juegos que coinciden con el término de búsqueda.
      operationId: searchGamesByName
      parameters:
        - name: name
          in: query
          required: true
          description: Nombre del juego a buscar
          schema:
            type: string
            maxLength: 255
      responses:
        '200':
          description: Búsqueda exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameDTO'
        '400':
          description: Parámetro 'name' no proporcionado

  /games/filter:
    post:
      tags:
        - Games
      summary: Filtrar videojuegos con criterios avanzados
      description: >
        Realiza una búsqueda avanzada de videojuegos en la API de IGDB utilizando una consulta de filtrado.
        Permite filtrar, ordenar y paginar los resultados.
      operationId: filterGames
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameFilterRequestDTO'
      responses:
        '200':
          description: Búsqueda avanzada exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameDTO'
        '400':
          description: Solicitud de filtrado inválida

  /games/{id}:
    get:
      tags:
        - Games
      summary: Obtener un videojuego por su ID
      description: >
        Obtiene los detalles completos de un videojuego específico a partir de su ID de IGDB.
      operationId: getGameById
      parameters:
        - name: id
          in: path
          required: true
          description: ID del juego a buscar
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Juego encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameDTO'
        '404':
          description: Juego no encontrado

  /platforms:
    get:
      tags:
        - Platforms
      summary: Listar todas las plataformas de videojuegos
      description: >
        Devuelve una lista de todas las plataformas de videojuegos disponibles en IGDB.
      operationId: listPlatforms
      responses:
        '200':
          description: Lista de plataformas obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlatformDTO'

  /users/register:
    post:
      tags:
        - Users
      summary: Registrar un nuevo usuario
      description: >
        Crea un nuevo usuario en el sistema.
        Requiere un nombre de usuario, un email y una contraseña.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequestDTO'
      responses:
        '201':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Datos de registro inválidos
        '409':
          description: El usuario con este email ya existe

  /users/login:
    post:
      tags:
        - Users
      summary: Iniciar sesión de usuario
      description: >
        Autentica a un usuario con su email y contraseña.
        Si las credenciales son válidas, devuelve los datos del usuario (sin la contraseña).
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDTO'
      responses:
        '200':
          description: Inicio de sesión exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDTO'
        '401':
          description: Credenciales inválidas

  /users/{userId}/games:
    get:
      tags:
        - Library
      summary: Listar todos los juegos en la biblioteca de un usuario
      description: >
        Devuelve una lista de todos los juegos que un usuario ha añadido a su biblioteca,
        incluyendo el estado de cada juego.
      operationId: listUserLibrary
      parameters:
        - name: userId
          in: path
          required: true
          description: ID del usuario
          schema:
            type: string
            format: uuid
            maxLength: 36
      responses:
        '200':
          description: Lista de juegos en la biblioteca del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserGameDTO'
        '404':
          description: Usuario no encontrado
    post:
      tags:
        - Library
      summary: Añadir un juego a la biblioteca de un usuario o actualizar su estado
      description: >
        Añade un nuevo juego a la biblioteca de un usuario o, si el juego ya existe,
        actualiza su estado (ej. de 'WANT_TO_PLAY' a 'PLAYING').
      operationId: addGameToLibrary
      parameters:
        - name: userId
          in: path
          required: true
          description: ID del usuario
          schema:
            type: string
            format: uuid
            maxLength: 36
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGameToLibraryRequestDTO'
      responses:
        '200':
          description: Juego añadido/actualizado en la biblioteca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGameDTO'
        '400':
          description: Solicitud inválida
        '404':
          description: Usuario o juego no encontrado

  /users/{userId}/games/{gameId}: # Nuevo endpoint para obtener el estado de un juego específico
    get:
      tags:
        - Library
      summary: Obtener el estado de un juego específico en la biblioteca del usuario
      description: >
        Devuelve el estado de un juego específico en la biblioteca del usuario.
        Si el juego no está en la biblioteca, devuelve 404.
      operationId: getUserGameStatus
      parameters:
        - name: userId
          in: path
          required: true
          description: ID del usuario
          schema:
            type: string
            format: uuid
            maxLength: 36
        - name: gameId
          in: path
          required: true
          description: ID del juego
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Estado del juego encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGameDTO'
        '404':
          description: Juego no encontrado en la biblioteca del usuario

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GameDTO:
      type: object
      properties:
        id:
          type: string
          description: El ID del juego (de IGDB).
        name:
          type: string
          description: El nombre del juego.
        genres:
          type: array
          items:
            type: string
          description: Lista de géneros del juego.
        releaseDate:
          type: string
          format: date
          description: Fecha de lanzamiento.
        coverImageUrl:
          type: string
          format: uri
          description: URL de la imagen de portada.
        summary:
          type: string
          description: Resumen detallado del juego.
        videos:
          type: array
          items:
            type: string
            format: uri
          description: Lista de URLs de videos relacionados.
        screenshots:
          type: array
          items:
            type: string
            format: uri
          description: Lista de URLs de capturas de pantalla.
        platforms:
          type: array
          items:
            type: string
          description: Lista de plataformas disponibles.
        rating:
          type: number
          format: double
          description: Puntuación media de los usuarios (0-100).

    PlatformDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: El ID de la plataforma (de IGDB).
        name:
          type: string
          description: El nombre de la plataforma.
        generation:
          type: integer
          format: int32
          description: La generación de la plataforma.
        platformType:
          type: string
          description: El tipo de plataforma.
          enum:
            - CONSOLE
            - ARCADE
            - PLATFORM
            - OPERATING_SYSTEM
            - PORTABLE_CONSOLE
            - COMPUTER
            - UNKNOWN

    UserDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email

    UserRegistrationRequestDTO:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    GameStatusDTO:
      type: string
      enum:
        - WANT_TO_PLAY
        - PLAYING
        - COMPLETED

    AddGameToLibraryRequestDTO:
      type: object
      required:
        - gameId
        - status
      properties:
        gameId:
          type: integer
          format: int64
          description: ID del juego a añadir (ID de IGDB).
        status:
          $ref: '#/components/schemas/GameStatusDTO'

    UserGameDTO:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        gameId:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/GameStatusDTO'
        addedAt:
          type: string
          format: date-time

    LoginRequestDTO:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Correo electrónico del usuario.
        password:
          type: string
          format: password
          description: Contraseña del usuario.

    LoginResponseDTO:
      type: object
      properties:
        token:
          type: string
          description: Token JWT para autenticación.
        userId:
          type: string
          format: uuid
          description: ID del usuario autenticado.

    GameFilterRequestDTO:
      type: object
      required:
        - filter
      properties:
        filter:
          type: string
          description: "La consulta de filtrado en el formato de la API de IGDB. Ejemplo: 'genres = (12, 32) & platforms = (6, 48)'"
        sort:
          type: string
          description: "El campo por el que ordenar los resultados. Ejemplo: 'total_rating desc'"
        limit:
          type: integer
          format: int32
          description: "El número máximo de resultados a devolver. Por defecto: 10."
          default: 10
        offset:
          type: integer
          format: int32
          description: "El número de resultados a saltar para la paginación. Por defecto: 0."
          default: 0